Move new list manipulation functionality to manipulateList.js
