+++ src/elements.js
@@ -1,6 +1,6 @@
 /* eslint-disable import/no-cycle */
-// import Interact from './interact.js';
 import Storage from './storage.js';
+import ListManipulation from './manipulateList.js';
 
 export default class Elmnts {
   static createCkBox(taskIndex, taskStatusDone, listToUpdate) {
@@ -11,12 +11,11 @@ export default class Elmnts {
     checkbox.addEventListener('change', () => {
       const updatedList = listToUpdate.changeStatusDone(taskIndex);
       Storage.saveAndUpdate(updatedList);
-      // Interact.populateUlTasksList(new ListOfTasks());
     });
     return checkbox;
   }
 
-  static createTaskDescr(taskIndex, taskDescription, listToUpdate) {
+  static createTaskDescr(taskIndex, taskDescription) {
     const description = document.createElement('p');
     description.contentEditable = true;
     description.style.outline = 'none';
@@ -25,9 +24,8 @@ export default class Elmnts {
     description.innerText = taskDescription;
     description.addEventListener('keypress', (e) => {
       if (e.key === 'Enter') {
-        const updatedList = listToUpdate.editDescription(taskIndex, e.target.innerText);
+        const updatedList = new ListManipulation().editDescription(taskIndex, e.target.innerText);
         Storage.saveAndUpdate(updatedList);
-        // Interact.populateUlTasksList(new ListOfTasks());
       }
     });
     return description;
@@ -52,7 +50,7 @@ export default class Elmnts {
     li.appendChild(descriptiveP);
     li.appendChild(icon);
     li.addEventListener('click', () => {
-      // reset style for all list items
+      // 1. reset style for all list items
       const allLis = document.querySelectorAll('.tasks_list > *');
       allLis.forEach((li) => {
         li.style.backgroundColor = 'white';
@@ -60,14 +58,15 @@ export default class Elmnts {
         li.children[2].style.cursor = 'move';
       });
 
-      // set style only for clicked
+      // 2. set style only for clicked
       li.style.backgroundColor = 'hsl(40deg 91% 75% / 75%)';
       icon.textContent = 'delete_forever';
       icon.style.cursor = 'pointer';
 
+      // 3. event Listener for remove Button
       icon.addEventListener('click', () => {
         if (icon.textContent === 'delete_forever') {
-          listToUpdate.removeByIndex(parseInt(icon.dataset.referTo, 10));
+          new ListManipulation().removeByIndex(parseInt(icon.dataset.referTo, 10));
         }
       });
     });
