+++ src/index.js
@@ -1,8 +1,9 @@
-import ListOfTasks from './list.js';
+// import ListOfTasks from './list.js';
+import ListManipulation from './manipulateList.js';
 import Interact from './interact.js';
 import './style.css';
 
-const tasks = new ListOfTasks();
+const tasks = new ListManipulation();
 window.onload = () => {
   Interact.populateUlTasksList(tasks);
 };
