+++ src/storage.js
@@ -3,10 +3,6 @@ import ListOfTasks from './list.js';
 import Interact from './interact.js';
 
 export default class Storage {
-  // static saveToStorage(objToStore, name = 'ToDoList') {
-  //   localStorage.setItem(name, JSON.stringify(objToStore));
-  // }
-
   static saveAndUpdate(objToStore, name = 'ToDoList') {
     localStorage.setItem(name, JSON.stringify(objToStore));
     Interact.populateUlTasksList(new ListOfTasks());
